fix(tinacms): scope the registerApi bridge guards to the in-flight event
A direction-wide boolean dropped any NEW event a listener dispatched
synchronously while a forward was mid-flight. The guards now compare the
event object itself, with save/restore so nesting works. The bridge tests
move to cms.test.ts, reviving the it.skip marker that has sat next to
registerApi since the api-to-cms direction first broke, and a new test
pins the nested-dispatch case (red-checked against the window-scoped
variant).
