+++ packages/tinacms/src/toolkit/core/cms.test.ts
@@ -74,8 +74,8 @@ describe('CMS', () => {
 
   describe('#registerApi', () => {
     describe('when the API has `events` of type EventBus', () => {
-      it.skip('events dispatched to the API are also sent through the CMS', () => {
-        const listener = jest.fn();
+      it('events dispatched to the API are also sent through the CMS', () => {
+        const listener = vi.fn();
         const event: CMSEvent = { type: 'api:example' };
         const example = { events: new EventBus() };
         const cms = new CMS({ apis: { example } });
@@ -84,6 +84,37 @@ describe('CMS', () => {
         cms.api.example.events.dispatch(event);
 
         expect(listener).toHaveBeenCalledWith(event);
+      });
+
+      it('forwards global-bus events to the api bus without recursing', () => {
+        const cms = new CMS();
+        const api = { events: new EventBus() };
+        cms.registerApi('thing', api);
+
+        const seen: CMSEvent[] = [];
+        api.events.subscribe('cms:did', (e) => seen.push(e));
+        // would stack-overflow if the two '*' bridges bounced the event back and forth
+        cms.events.dispatch({ type: 'cms:did' });
+
+        expect(seen).toHaveLength(1);
+      });
+
+      it('bridges a new event dispatched synchronously by a listener mid-forward', () => {
+        const cms = new CMS();
+        const api = { events: new EventBus() };
+        cms.registerApi('thing', api);
+
+        const seenOnApi: CMSEvent[] = [];
+        api.events.subscribe('secondary:event', (e) => seenOnApi.push(e));
+        // reacting to the bridged event by dispatching a different one must not
+        // be swallowed by the re-entrancy guard
+        cms.events.subscribe('thing:did', () => {
+          cms.events.dispatch({ type: 'secondary:event' });
+        });
+
+        api.events.dispatch({ type: 'thing:did' });
+
+        expect(seenOnApi).toHaveLength(1);
       });
     });
   });
